Fully working prime factorization app.
We're currently working on developing a Service Weaver workshop, where
participants can walk through writing a Service Weaver application from
scratch. Our idea is to have a simple prime factorization application
that shows off components, logging, metrics, testing, and routing. The
application can also be extended to compare the calculated prime
factorizations to the output of ChatGPT, to give the app a bit of spice.

This commit introduces a fully working initial version of the
application. In future commits, I'll split the application into steps
with instructions.