Cancel stale operations to prevent UI races.
When a user clicks the "Factor" button, it sends a request to the
backend to factor the provided number. When the backend returns, the
frontend updates the UI with the factors.

Before this PR, this logic was susceptible to races. In particular, if a
user requests the factorization or a large prime, the request lingers
for a while. If they give up and request the factorization of a much
smaller number, the factorization finishes very quickly. Then, the
lingering factorization finishes and overwrites the current
factorization. This is bad.

This PR fixes the bug. Now, whenever a user clicks the "Factor" button,
the previous request is cancelled, if it didn't finish already.
Relatedly, I also updated the Factorer component to respect cancelled
contexts.