Added workshop instructions (except ChatGPT).
This PR adds the instructions that participants will follow (except for
the ChatGPT part which I'll finish in a future PR). Once we finalize the
instructions, I'll split the solutions into the various directories and
embed references to them in the instructions. I'm also thinking of
adding some quiz questions to test the participants' understanding of
things. I'll also add a link to a Google Form where people can leave
feedback.