Added unit tests for instructions.
This PR adds a set of unit tests that run the commands listed in the
workshop's instructions. Note that the unit tests literally run the
commands, rather than doing something more direct, like the tests in
searcher_test.go. This is intentional and aims to catch any small errors
in the instructions we give participants.