Fixed unkilled process bug in instructions tests.
Some of the instructions tests run "go run ." in a subprocess and kill
the subprocess at the end of the test. However, "go run ." internally
builds a binary and runs it as a subprocess. Killing "go run ." was
killing "go run ." (duh), but it wasn't killing the underlying binary.
This PR fixes that by building and running the binary directly, rather
than using "go run .".