Fixed caching of queries with no matching emojis.
Recall that the `Searcher` component has a `Search` method that returns
the set of emojis that match a provided query. In part 7 of the
workshop, we add a `Cache` component and update the `Searcher` component
to cache query results.

@spetrovic77 pointed out that our caching is buggy. Specifically, we do
not cache queries that don't have any matching emojis. This PR fixes the
bug, being careful to distinguish between when the `Cache` returns `nil`
(meaning there is nothing cached) and when it returns `[]string{}`
(meaning it has cached the fact that there are no matching emojis).

A better solution would be to introduce a custom error value that is
returned when there is a cache miss, but that would require explaining
`AutoMarshal` and error serialization, which I think would be too
complicated for the workshop. Instead, we do a slightly jankier but
simpler fix.